Add flex container and properties to the design
Add flex container to body and direction to column.
Remove dots on the list items and make them flex container.
Make header and footer flex container with space-between.
Add background color to footer.
Make content flex container with direction as column and center it.
Add padding to button.
Remove the decorative line under the links and make the color gray.
Add enough gap to list and content and padding to header and footer.